+++ source/braille.py
@@ -66,7 +66,7 @@ import queueHandler
 import brailleViewer
 from autoSettingsUtils.driverSetting import BooleanDriverSetting, NumericDriverSetting
 from utils.security import objectBelowLockScreenAndWindowsIsLocked, post_sessionLockStateChanged
-from textUtils import isUnicodeNormalized, UnicodeNormalizationOffsetConverter
+from textUtils import isUnicodeNormalized, OffsetConverter, UnicodeNormalizationOffsetConverter
 import hwIo
 from editableText import EditableText
 
@@ -568,10 +568,17 @@ class Region(object):
 		if config.conf["braille"]["expandAtCursor"] and self.cursorPos is not None:
 			mode |= louis.compbrlAtCursor
 
-		converter: UnicodeNormalizationOffsetConverter | None = None
+		converter: OffsetConverter | None = None
 		textToTranslate = self.rawText
 		textToTranslateTypeforms = self.rawTextTypeforms
 		cursorPos = self.cursorPos
+		if config.conf["braille"]["translationTable"].startswith("zh"):
+			from textUtils.wordSeg.wordSegUtils import WordSegWithSeparatorOffsetConverter  # noqa: F401
+
+			converter = WordSegWithSeparatorOffsetConverter(textToTranslate)
+			textToTranslate = converter.encoded
+			if cursorPos is not None:
+				cursorPos = converter.strToEncodedOffsets(cursorPos)
 		if config.conf["braille"]["unicodeNormalization"] and not isUnicodeNormalized(textToTranslate):
 			converter = UnicodeNormalizationOffsetConverter(textToTranslate)
 			textToTranslate = converter.encoded